start csees.fi problems

+++ 01_introductory_problems/missing_number.cpp
@@ -1,5 +1,4 @@
 #include <bits/stdc++.h>
-#include <iostream>
 
 using namespace std;
 
@@ -8,7 +7,7 @@ int main() {
   ios::sync_with_stdio(0);
   cin.tie(0);
 
-  long long n;
+  int n;
   cin >> n;
   while (true) {
     cout << n << " ";
